added try function and error handling
